updates for 2.x

+++ Controller/CronTasksController.php
@@ -30,11 +30,11 @@ class CronTasksController extends QueueAppController {
 	}
 
 	public function add() {
-		if (!empty($this->request->data)) {
+		if ($this->Common->isPosted()) {
 			$this->CronTask->create();
 			if ($this->CronTask->save($this->request->data)) {
 				$var = $this->request->data['CronTask']['title'];
-				$this->Common->flashMessage(sprintf(__('record add %s saved'), h($var)), 'success');
+				$this->Common->flashMessage(__('record add %s saved', h($var)), 'success');
 				$this->Common->postRedirect(array('action' => 'index'));
 			} else {
 				$this->Common->flashMessage(__('formContainsErrors'), 'error');
@@ -47,10 +47,10 @@ class CronTasksController extends QueueAppController {
 			$this->Common->flashMessage(__('invalid record'), 'error');
 			$this->Common->autoRedirect(array('action' => 'index'));
 		}
-		if (!empty($this->request->data)) {
+		if ($this->Common->isPosted()) {
 			if ($this->CronTask->save($this->request->data)) {
 				$var = $this->request->data['CronTask']['title'];
-				$this->Common->flashMessage(sprintf(__('record edit %s saved'), h($var)), 'success');
+				$this->Common->flashMessage(__('record edit %s saved', h($var)), 'success');
 				$this->Common->postRedirect(array('action' => 'index'));
 			} else {
 				$this->Common->flashMessage(__('formContainsErrors'), 'error');
@@ -62,17 +62,20 @@ class CronTasksController extends QueueAppController {
 	}
 
 	public function delete($id = null) {
+		if (!$this->Common->isPosted()) {
+			throw new MethodNotAllowedException();
+		}
 		if (empty($id) || !($cronTask = $this->CronTask->find('first', array('conditions'=>array('CronTask.id'=>$id), 'fields'=>array('id', 'title'))))) {
 			$this->Common->flashMessage(__('invalid record'), 'error');
 			$this->Common->autoRedirect(array('action'=>'index'));
 		}
 		$var = $cronTask['CronTask']['title'];
-		
+
 		if ($this->CronTask->delete($id)) {
-			$this->Common->flashMessage(sprintf(__('record del %s done'), h($var)), 'success');
+			$this->Common->flashMessage(__('record del %s done', h($var)), 'success');
 			$this->redirect(array('action' => 'index'));
 		}
-		$this->Common->flashMessage(sprintf(__('record del %s not done exception'), h($var)), 'error');
+		$this->Common->flashMessage(__('record del %s not done exception', h($var)), 'error');
 		$this->Common->autoRedirect(array('action' => 'index'));
 	}
 
@@ -97,11 +100,11 @@ class CronTasksController extends QueueAppController {
 	}
 
 	public function admin_add() {
-		if (!empty($this->request->data)) {
+		if ($this->Common->isPosted()) {
 			$this->CronTask->create();
 			if ($this->CronTask->save($this->request->data)) {
 				$var = $this->request->data['CronTask']['title'];
-				$this->Common->flashMessage(sprintf(__('record add %s saved'), h($var)), 'success');
+				$this->Common->flashMessage(__('record add %s saved', h($var)), 'success');
 				$this->Common->postRedirect(array('action' => 'index'));
 			} else {
 				$this->Common->flashMessage(__('formContainsErrors'), 'error');
@@ -114,10 +117,10 @@ class CronTasksController extends QueueAppController {
 			$this->Common->flashMessage(__('invalid record'), 'error');
 			$this->Common->autoRedirect(array('action' => 'index'));
 		}
-		if (!empty($this->request->data)) {
+		if ($this->Common->isPosted()) {
 			if ($this->CronTask->save($this->request->data)) {
 				$var = $this->request->data['CronTask']['title'];
-				$this->Common->flashMessage(sprintf(__('record edit %s saved'), h($var)), 'success');
+				$this->Common->flashMessage(__('record edit %s saved', h($var)), 'success');
 				$this->Common->postRedirect(array('action' => 'index'));
 			} else {
 				$this->Common->flashMessage(__('formContainsErrors'), 'error');
@@ -129,17 +132,20 @@ class CronTasksController extends QueueAppController {
 	}
 
 	public function admin_delete($id = null) {
+		if (!$this->Common->isPosted()) {
+			throw new MethodNotAllowedException();
+		}
 		if (empty($id) || !($cronTask = $this->CronTask->find('first', array('conditions'=>array('CronTask.id'=>$id), 'fields'=>array('id', 'title'))))) {
 			$this->Common->flashMessage(__('invalid record'), 'error');
 			$this->Common->autoRedirect(array('action'=>'index'));
 		}
 		$var = $cronTask['CronTask']['title'];
-		
+
 		if ($this->CronTask->delete($id)) {
-			$this->Common->flashMessage(sprintf(__('record del %s done'), h($var)), 'success');
+			$this->Common->flashMessage(__('record del %s done', h($var)), 'success');
 			$this->redirect(array('action' => 'index'));
 		}
-		$this->Common->flashMessage(sprintf(__('record del %s not done exception'), h($var)), 'error');
+		$this->Common->flashMessage(__('record del %s not done exception', h($var)), 'error');
 		$this->Common->autoRedirect(array('action' => 'index'));
 	}
